chapter-08/simple-chat-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.51.0
python-dotenv==1.0.0
httpx==0.27.0
